+++ packages/reactivity/src/reactive/index.ts
@@ -1,0 +1,4 @@
+export { makeMapReactive } from './collectionMethods';
+export { proxyMap } from './proxyMap';
+export { reactive, wrap, toRaw } from './reactive';
+export { reactiveNodes } from './reactiveNodes';
