+++ packages/reactivity/src/reactive/proxyMap.ts
@@ -1,0 +1,1 @@
+export const proxyMap = new WeakMap<object, object>();
