+++ packages/reactivity/src/reactive/reactiveNodes.ts
@@ -1,0 +1,3 @@
+import { Signal } from '../Signal';
+
+export const reactiveNodes = new WeakMap<object, Map<unknown, Signal>>();
