+++ packages/reactivity/src/reactive/symbols.ts
@@ -1,0 +1,2 @@
+export const $PROXY = Symbol('$PROXY');
+export const $RAW = Symbol('$RAW');
